Build with -Wextra and -Werror=format-security
We'll need to suppress a few warnings, and fix some code that triggers
-Wsign-compare.

One -Wsign-compare is not fixed here because it's a real bug, to be
addressed separately.

We'll also manually add -Werror=format-security since this is amazingly
not included in -Wextra.

+++ lib/widgets/ephy-location-entry.c
@@ -725,7 +725,7 @@ button_box_size_allocated_cb (GtkWidget    *widget,
   GtkStateFlags state_flags;
   GtkBorder padding;
 
-  if (lentry->allocation_width == allocation->width)
+  if (lentry->allocation_width == (guint)allocation->width)
     return;
 
   lentry->allocation_width = allocation->width;
